refactor HiltExt into a custom test rule that cleans up activity scenarios
