+++ app/src/test/java/org/groundplatform/android/ui/offlineareas/selector/OfflineAreaSelectorFragmentTest.kt
@@ -35,8 +35,8 @@ import junit.framework.Assert.assertFalse
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.test.advanceUntilIdle
 import org.groundplatform.android.BaseHiltTest
+import org.groundplatform.android.FragmentScenarioRule
 import org.groundplatform.android.R
-import org.groundplatform.android.launchFragmentInHiltContainer
 import org.groundplatform.android.repository.OfflineAreaRepository
 import org.junit.Assert.assertNull
 import org.junit.Before
@@ -58,11 +58,12 @@ class OfflineAreaSelectorFragmentTest : BaseHiltTest() {
   private val offlineAreaRepository: OfflineAreaRepository = mock()
 
   @get:Rule override val composeTestRule = createAndroidComposeRule<ComponentActivity>()
+  @get:Rule val fragmentScenario = FragmentScenarioRule()
 
   @Before
   override fun setUp() {
     super.setUp()
-    launchFragmentInHiltContainer<OfflineAreaSelectorFragment> {
+    fragmentScenario.launchFragmentInHiltContainer<OfflineAreaSelectorFragment> {
       fragment = this as OfflineAreaSelectorFragment
     }
   }
