+++ app/src/test/java/org/groundplatform/android/ui/tos/TermsOfServiceFragmentTest.kt
@@ -35,10 +35,9 @@ import javax.inject.Inject
 import kotlin.test.assertEquals
 import kotlin.test.assertNull
 import org.groundplatform.android.BaseHiltTest
+import org.groundplatform.android.FragmentScenarioRule
 import org.groundplatform.android.R
 import org.groundplatform.android.data.remote.FakeRemoteDataStore
-import org.groundplatform.android.launchFragmentInHiltContainer
-import org.groundplatform.android.launchFragmentWithNavController
 import org.groundplatform.android.model.TermsOfService
 import org.groundplatform.android.repository.TermsOfServiceRepository
 import org.groundplatform.android.system.NetworkManager
@@ -53,6 +52,7 @@ import org.robolectric.RobolectricTestRunner
 @RunWith(RobolectricTestRunner::class)
 @Suppress("StringShouldBeRawString")
 class TermsOfServiceFragmentTest : BaseHiltTest() {
+  @get:Rule val fragmentScenario = FragmentScenarioRule()
 
   @Inject lateinit var fakeRemoteDataStore: FakeRemoteDataStore
   @Inject lateinit var termsOfServiceRepository: TermsOfServiceRepository
@@ -74,7 +74,9 @@ class TermsOfServiceFragmentTest : BaseHiltTest() {
 
   @Test
   fun `Toolbar is displayed`() {
-    launchFragmentInHiltContainer<TermsOfServiceFragment>(bundleOf(Pair("isViewOnly", false)))
+    fragmentScenario.launchFragmentInHiltContainer<TermsOfServiceFragment>(
+      bundleOf(Pair("isViewOnly", false))
+    )
 
     composeTestRule
       .onNodeWithText(composeTestRule.activity.getString(R.string.tos_title))
@@ -83,14 +85,18 @@ class TermsOfServiceFragmentTest : BaseHiltTest() {
 
   @Test
   fun `Toolbar Back Arrow is displayed`() {
-    launchFragmentInHiltContainer<TermsOfServiceFragment>(bundleOf(Pair("isViewOnly", true)))
+    fragmentScenario.launchFragmentInHiltContainer<TermsOfServiceFragment>(
+      bundleOf(Pair("isViewOnly", true))
+    )
 
     composeTestRule.onNodeWithContentDescription("Back").assertIsDisplayed()
   }
 
   @Test
   fun `Terms of service text should be displayed`() {
-    launchFragmentInHiltContainer<TermsOfServiceFragment>(bundleOf(Pair("isViewOnly", false)))
+    fragmentScenario.launchFragmentInHiltContainer<TermsOfServiceFragment>(
+      bundleOf(Pair("isViewOnly", false))
+    )
 
     composeTestRule.onNodeWithText("This is a heading\n\nSample terms of service\n\n").isDisplayed()
 
@@ -104,7 +110,9 @@ class TermsOfServiceFragmentTest : BaseHiltTest() {
 
   @Test
   fun `Agree button should not be enabled by default`() {
-    launchFragmentInHiltContainer<TermsOfServiceFragment>(bundleOf(Pair("isViewOnly", false)))
+    fragmentScenario.launchFragmentInHiltContainer<TermsOfServiceFragment>(
+      bundleOf(Pair("isViewOnly", false))
+    )
 
     getCheckbox().assertIsDisplayed()
     getButton().assertIsNotEnabled()
@@ -112,7 +120,9 @@ class TermsOfServiceFragmentTest : BaseHiltTest() {
 
   @Test
   fun `Agree button should be enabled when checkbox is clicked`() {
-    launchFragmentInHiltContainer<TermsOfServiceFragment>(bundleOf(Pair("isViewOnly", false)))
+    fragmentScenario.launchFragmentInHiltContainer<TermsOfServiceFragment>(
+      bundleOf(Pair("isViewOnly", false))
+    )
 
     getCheckbox().performClick()
 
@@ -121,7 +131,7 @@ class TermsOfServiceFragmentTest : BaseHiltTest() {
 
   @Test
   fun `Agree button should update preferences and navigate when pressed`() = runWithTestDispatcher {
-    launchFragmentWithNavController<TermsOfServiceFragment>(
+    fragmentScenario.launchFragmentWithNavController<TermsOfServiceFragment>(
       fragmentArgs = bundleOf(Pair("isViewOnly", false)),
       destId = R.id.terms_of_service_fragment,
       navControllerCallback = { navController = it },
@@ -138,7 +148,9 @@ class TermsOfServiceFragmentTest : BaseHiltTest() {
 
   @Test
   fun `View only mode hides controls`() = runWithTestDispatcher {
-    launchFragmentInHiltContainer<TermsOfServiceFragment>(bundleOf(Pair("isViewOnly", true)))
+    fragmentScenario.launchFragmentInHiltContainer<TermsOfServiceFragment>(
+      bundleOf(Pair("isViewOnly", true))
+    )
 
     getCheckbox().assertIsNotDisplayed()
     getButton().assertIsNotDisplayed()
@@ -147,7 +159,7 @@ class TermsOfServiceFragmentTest : BaseHiltTest() {
   @Test
   fun `Open SurveySelectorFragment with survey ID`() = runWithTestDispatcher {
     val uri = Uri.parse("https://groundplatform.org/android/survey/surveyId")
-    launchFragmentWithNavController<TermsOfServiceFragment>(
+    fragmentScenario.launchFragmentWithNavController<TermsOfServiceFragment>(
       fragmentArgs = bundleOf("isViewOnly" to false),
       destId = R.id.terms_of_service_fragment,
       navControllerCallback = { navController = it },
@@ -161,7 +173,7 @@ class TermsOfServiceFragmentTest : BaseHiltTest() {
 
   @Test
   fun `Open SurveySelectorFragment without survey ID`() = runWithTestDispatcher {
-    launchFragmentWithNavController<TermsOfServiceFragment>(
+    fragmentScenario.launchFragmentWithNavController<TermsOfServiceFragment>(
       fragmentArgs = bundleOf("isViewOnly" to false),
       destId = R.id.terms_of_service_fragment,
       navControllerCallback = { navController = it },
